requirements.txt
discord.py
dblpy
pymongo
python-dotenv
humanfriendly
emoji
google-api-python-client